% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAcknowledgement.R
\name{addAcknowledgement}
\alias{addAcknowledgement}
\title{addAcknowledgement}
\usage{
addAcknowledgement(theData = NULL, apiKey, acknowledgement_source)
}
\arguments{
\item{theData}{Data frame of the data returned from getPublication.R}

\item{apiKey}{Character string with your Scopus API key}

\item{acknowledgement_source}{List of organisations and the keywords to be searched signalling their acknowledgement}
}
\value{
Data frame of publications returned by search query
}
\description{
Function used to identify publications that have acknowledged a specified organisation.
}
\examples{

acknowledgement_search <- list("ANTNZ" = 'FUND-ALL("Antarctica New Zealand" OR "Antarctica NZ") AND PUBYEAR > 2021')

addAcknowledgement(theData = df,
                   theApiKey = your_api_key,
                   acknowledgement_source = acknowledgement_search)


}
