% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAuthorMetadata.R
\name{getAuthorMetadata}
\alias{getAuthorMetadata}
\title{getAuthorMetadata}
\usage{
getAuthorMetadata(theScopusID, theApiKey)
}
\arguments{
\item{theScopusID}{Character string with the publications Scopus ID}

\item{theApiKey}{Character string with your API key}

\item{theDOI}{Character string with the publication's doi}

\item{start}{Numeric with the first citation year required}

\item{end}{Numeric with the final citation year required}
}
\value{
Citation data with the input identifier
}
\description{
Function that uses the publication DOI or Scopus ID to get citations for a publication
}
\examples{
getAuthorMetadata(theScopusID = 8044727400,
            theApiKey = your_api_key)

}
