% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCitation.R
\name{getCitation}
\alias{getCitation}
\title{getCitation}
\usage{
getCitation(theDOI = NULL, theScopusID = NULL, start, end, theApiKey)
}
\arguments{
\item{theDOI}{Character string with the publication's doi}

\item{theScopusID}{Character string with the publications Scopus ID}

\item{start}{Numeric with the first citation year required}

\item{end}{Numeric with the final citation year required}

\item{theApiKey}{Character string with your API key}
}
\value{
Citation data with the input identifier
}
\description{
Function that uses the publication DOI or Scopus ID to get citations for a publication
}
\examples{
getCitation(theDOI = "10.1080/16000889.2021.1933783",
            start = 2021,
            end = 2022,
            theApiKey = your_api_key)


}
