% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCiteScore.R
\name{getCiteScore}
\alias{getCiteScore}
\title{getCiteScore}
\usage{
getCiteScore(theScopusID, theApiKey)
}
\arguments{
\item{theScopusID}{Character string with the publication's scopus id}

\item{theApiKey}{Character string with your API key}
}
\value{
Associated journal's CiteScore
}
\description{
Function that uses the publication DOI or Scopus ID to get citations for a publication
}
\examples{
getCiteScore(theScopusID = "85101941134",
             theApiKey = your_api_key)


}
