% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{getMetadata}
\alias{getMetadata}
\title{getMetadata}
\usage{
getMetadata(theScopusID, theApiKey, subjectMetadata)
}
\arguments{
\item{theScopusID}{Character string with the Scopus ID}

\item{theApiKey}{Character string with your Scopus API key}
}
\value{
Data frame of the metadata and the Scopus ID
}
\description{
Function used to get author and affiliation metadata for a publication record
}
\examples{
getMetadata(theScopusID = "85103423955",
            theApiKey = your_api_key)


}
