% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOnlineDate.R
\name{getOnlineDate}
\alias{getOnlineDate}
\title{getOnlineDate}
\usage{
getOnlineDate(theDOI = NULL, theEID = NULL, theApiKey)
}
\arguments{
\item{theDOI}{Character string with the publication's doi}

\item{theEID}{Character string with the publication's EID}

\item{theApiKey}{Character string with your API key}
}
\value{
Data frame with the online date and the input identifier
}
\description{
Function that gets the first available online date for Science Direct publications
}
\examples{
getCitation(theDOI = "10.1080/16000889.2021.1933783",
            start = 2021,
            end = 2022,
            theApiKey = your_api_key)


}
