% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPercentile.R
\name{getPercentile}
\alias{getPercentile}
\title{getPercentile}
\usage{
getPercentile(thePublicationID, yearRange, theApiKey)
}
\arguments{
\item{thePublicationID}{Character string with the journal's source id/publication id}

\item{yearRange}{The years to be returned. Possible values: 3, 5, 10.}

\item{theApiKey}{Character string with your API key}
}
\value{
Associated journal's percentiles (1st, 5th, 10th, 25th)
}
\description{
Function that uses the publication DOI or Scopus ID to get citations for a publication
}
\examples{
getPercentiles(thePublicationID = "21100790929",
               yearRange = "3",   
               theApiKey = your_api_key)


}
