% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPublication.R
\name{getPublication}
\alias{getPublication}
\title{getPublication}
\usage{
getPublication(theQuery = NULL, theApiKey)
}
\arguments{
\item{theQuery}{Character string with the Scopus query}

\item{apiKey}{Character string with your Scopus API key}
}
\value{
Data frame of publications returned by search query
}
\description{
Function used to pull publication records from Scopus Search API
}
\examples{
getPublication(theQuery = 'TITLE-ABS-KEY("Scott Base) AND PUBYEAR IS 2021',
               theApiKey = your_api_key)


}
