% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScienceDirect.R
\name{getScienceDirect}
\alias{getScienceDirect}
\title{getScienceDirect}
\usage{
getScienceDirect(theQuery = NULL, theApiKey)
}
\arguments{
\item{theQuery}{Character string with the ScienceDirect query}

\item{apiKey}{Character string with your Scopus API key}
}
\value{
Data frame of publications returned by search query
}
\description{
Function used to pull publication records from ScienceDirect API
}
\examples{
getScienceDirect(theQuery = 'doi(10.1016/B978-0-12-819109-5.00008-6)',
                 theApiKey = your_api_key)


}
