% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubject.R
\name{getSubject}
\alias{getSubject}
\title{getSubject}
\usage{
getSubject(abstractData = NULL, theScopusID = NULL, theApiKey)
}
\arguments{
\item{abstractData}{Data frame of data prepared by getMetadata.R}

\item{theScopusID}{Character string with the publications Scopus ID}

\item{theApiKey}{Character string with your API key}
}
\value{
Subject metadata with the scoups id
}
\description{
Function that uses data returned from abstract API within getMetadata to get subject metadata
}
