% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAntarcticDatabase.R
\name{makeAntarcticDatabase}
\alias{makeAntarcticDatabase}
\title{makeAntarcticDatabase}
\usage{
makeAntarcticDatabase(apiKey)
}
\arguments{
\item{apiKey}{Character string with your Scopus API key}
}
\value{
Data frame of NZ Antarctic Science publications
}
\description{
High level wrapper function which builds NZ Antarctic Database
}
\examples{

nzAntScience <- makeAntarcticDatabase(apiKey = yourApiKey)

}
