% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePublicationTable.R
\name{makePublicationTable}
\alias{makePublicationTable}
\title{makePublicationTable}
\usage{
makePublicationTable(
  theQuery = NULL,
  apiKey = NULL,
  acknowledgement_source = NULL,
  citations = FALSE,
  articleMetadata = FALSE,
  onlineDate = FALSE,
  subjectMetadata = TRUE,
  metrics = FALSE,
  percentileYearRange = 3
)
}
\arguments{
\item{theQuery}{Character string with the Scopus query}

\item{apiKey}{Character string with your Scopus API key}

\item{acknowledgement_source}{List which includes keywords used to search publication acknowledgement section}

\item{citations}{Logical to attach citations to search results}

\item{articleMetadata}{Logical to attach author, affiliation and subject metadata}

\item{onlineDate}{Logical to get first available online date for publications in Science Direct}
}
\value{
Data frame of publications returned by search query
}
\description{
High level wrapper function which aggregates metadata from each Scopus API
}
\examples{
query = '(TITLE-ABS-KEY(antarct* OR "southern ocean" OR "ross sea" OR "amundsen sea" OR "weddell sea") AND NOT TITLE-ABS-KEY(candida OR "except antarctica" OR "not antarctica")) AND PUBYEAR > 2021 AND AFFILCOUNTRY("New Zealand")'

acknowledgement_search <- list("ANTNZ" = 'FUND-ALL("Antarctica New Zealand" OR "Antarctica NZ") AND PUBYEAR > 2021',
                               "ASP" = 'FUND-ALL(Antarctic AND Science AND Platform OR ASP OR "ANTA 1801") AND PUBYEAR > 2021',
                               "NZARI" = 'FUND-ALL("NZARI" OR "New Zealand Antarctic Research Institute") AND PUBYEAR > 2021')

makePublicationTable(theQuery = query,
                     apiKey = your_api_key,
                     citations = TRUE,
                     articleMetadata = TRUE,
                     acknowledgement_source = acknowledgement_search)

}
