% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRelationalDatabase.R
\name{makeRelationalDatabase}
\alias{makeRelationalDatabase}
\title{makeRelationalDatabase}
\usage{
makeRelationalDatabase(theData)
}
\arguments{
\item{theData}{Data frame returned by makePublicationTable.R}
}
\value{
List with 3 elements
}
\description{
Function used to convert data from makePublicationTable.R into a relational database
}
\examples{
makeRelationalDatabase(theData = theData)

}
